split off functionality from lib.rs into apply_points.rs
